% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitfngr.R
\name{fngr}
\alias{fngr}
\title{Access a list of values separately but calculate them together.}
\usage{
fngr(func, check_all = FALSE, recalculate_indices = 1)
}
\arguments{
\item{func}{Function that returns a list of values}

\item{check_all}{Should it check that the accessed values were
calculated at the current input?}

\item{recalculate_indices}{Indices for which the values should
be recalculated.}
}
\value{
An environment where the function values are calculated.
}
\description{
Access a list of values separately but calculate them together.
}
\examples{
tfunc <- function(x) {list(x, x+1, x+2, x+3, x+4)}
tenv <- fngr(tfunc)
tenv$f(1)(0)
tenv$f(3)(0)
tenv$f(3)(1)
tenv$f(1)(23.4)
tenv$f(4)()
}

